changed to double weights

+++ src/caseloader/Problem.java
@@ -5,7 +5,7 @@ public interface Problem {
 	
 	public int getInputSize();
 	
-	public int[] getCase(int i);
+	public double[] getCase(int i);
 	
 	public int[] getPrefDim();
